requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
requests
pystray
Pillow
pytest
pytest-playwright
winreg; sys_platform == 'win32'
cryptography
httpx
slowapi
